Add custom website feature - Add Edit component which contains the form for custom website - Add View component which displays the custom website - Configure url updates by using the useQueryParam hook in Custom component   and passing the functions supplied by the hook to update params to Edit   and view components.
The custom component checks if there are any query parameters supplied
to it and makes necessary changes in the redux state.
This process happens once. (see useEffect hook in custom component)

All the input controls in the Edit component trigger query and redux state
changes.
mainImage and bgImage query params are now based on ArrayParam type
supplied by useQueryParam.
The conversion of comma separated string to query params is done in
customWebsiteDataSlice.js